Use prebuilt nginx package with geoip.
This commit adds a pre-built nginx debian package to the repository
and installs nginx using that package.  Ideally, this package would be
stored on an asset server or apt repository.

Despite the loss of history, we might consider a git-filter-branch
after we are able to remove this asset from the repo.